+++ web/modules/custom/filmfolk_fixtures/src/Fixture/PersonFixture.php
@@ -2,30 +2,15 @@
 
 namespace Drupal\filmfolk_fixtures\Fixture;
 
-use Drupal\content_fixtures\Fixture\AbstractFixture;
 use Drupal\content_fixtures\Fixture\DependentFixtureInterface;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\filmfolk\Plugin\Field\FieldType\FunktionErfaringItem;
-use Drupal\user\Entity\User;
 use Drupal\user\UserInterface;
-use Drupal\user\UserStorageInterface;
 
 /**
- * User fixture.
+ * Person fixture.
  */
-final class PersonFixture extends AbstractFixture implements DependentFixtureInterface {
-  private const ROLE_PERSON = 'person';
-
-  /**
-   * The user storage.
-   */
-  private readonly UserStorageInterface $userStorage;
-
-  public function __construct(
-    EntityTypeManagerInterface $entityTypeManager,
-  ) {
-    $this->userStorage = $entityTypeManager->getStorage('user');
-  }
+final class PersonFixture extends UserFixture implements DependentFixtureInterface {
+  const ROLE_PERSON_ID = 'person';
 
   /**
    * {@inheritdoc}
@@ -36,8 +21,8 @@ final class PersonFixture extends AbstractFixture implements DependentFixtureInt
     // Get the pre-created media entity using the reference key we set.
     $default_profile_picture_media = $this->getReference('media:profile_picture:default');
 
-    $person = $this->createPerson([
-      'mail' => 'user@example.com',
+    $user = $this->createUser([
+      'mail' => 'person@example.com',
       'field_navn' => 'Navn Navnesen',
       'field_kommune' => $this->getReference('kommune:Aarhus'),
       'field_funktion' => [
@@ -54,10 +39,10 @@ final class PersonFixture extends AbstractFixture implements DependentFixtureInt
       ],
     ])
       ->activate();
-    $person->save();
+    $user->save();
 
-    $person = $this->userStorage->create([
-      'mail' => 'user002@example.com',
+    $user = $this->createUser([
+      'mail' => 'person002@example.com',
       'field_navn' => 'Navn Navnesen',
       'field_kommune' => $this->getReference('kommune:Herning'),
       'field_funktion' => [
@@ -81,10 +66,10 @@ final class PersonFixture extends AbstractFixture implements DependentFixtureInt
       ],
     ])
       ->activate();
-    $person->save();
+    $user->save();
 
-    $person = User::create([
-      'mail' => 'user003@example.com',
+    $user = $this->createUser([
+      'mail' => 'person003@example.com',
       'field_navn' => 'Navn Navnesen',
       'field_kommune' => $this->getReference('kommune:Herning'),
       'field_funktion' => [
@@ -108,10 +93,10 @@ final class PersonFixture extends AbstractFixture implements DependentFixtureInt
       ],
     ])
       ->activate();
-    $person->save();
+    $user->save();
 
-    $person = $this->createPerson([
-      'mail' => 'user004@example.com',
+    $user = $this->createUser([
+      'mail' => 'person004@example.com',
       'field_navn' => 'Navn Navnesen',
       'field_kommune' => $this->getReference('kommune:Herning'),
       'field_funktion' => [
@@ -135,10 +120,10 @@ final class PersonFixture extends AbstractFixture implements DependentFixtureInt
       ],
     ])
       ->activate();
-    $person->save();
+    $user->save();
 
-    $person = $this->createPerson([
-      'mail' => 'user005@example.com',
+    $user = $this->createUser([
+      'mail' => 'person005@example.com',
       'field_navn' => 'Navn Navnesen',
       'field_kommune' => $this->getReference('kommune:Herning'),
       'field_funktion' => [
@@ -162,10 +147,10 @@ final class PersonFixture extends AbstractFixture implements DependentFixtureInt
       ],
     ])
       ->activate();
-    $person->save();
+    $user->save();
 
-    $person = $this->createPerson([
-      'mail' => 'user006@example.com',
+    $user = $this->createUser([
+      'mail' => 'person006@example.com',
       'field_navn' => 'Navn Navnesen',
       'field_kommune' => $this->getReference('kommune:Herning'),
       'field_funktion' => [
@@ -189,21 +174,21 @@ final class PersonFixture extends AbstractFixture implements DependentFixtureInt
       ],
     ])
       ->activate();
-    $person->save();
+    $user->save();
 
-    $person = $this->createPerson([
-      'mail' => 'user1@example.com',
+    $user = $this->createUser([
+      'mail' => 'person1@example.com',
       'field_kommune' => $this->getReference('kommune:Aarhus'),
     ]);
-    $person->save();
+    $user->save();
   }
 
   /**
-   * Create a person.
+   * {@inheritdoc}
    */
-  private function createPerson(array $values): UserInterface {
-    return $this->userStorage->create($values)
-      ->addRole(self::ROLE_PERSON);
+  protected function createUser(array $values = []): UserInterface {
+    return parent::createUser($values)
+      ->addRole(self::ROLE_PERSON_ID);
   }
 
   /**